Move Delete key to diagram signal
Instead of a tool.
