+++ gaphor/ui/diagrampage.py
@@ -10,8 +10,10 @@ from gaphas.tool.rubberband import RubberbandPainter, RubberbandState
 from gaphas.view import GtkView
 from gi.repository import Adw, Gdk, GdkPixbuf, Gio, GLib, Gtk
 
+from gaphor import UML
 from gaphor.core import event_handler, gettext
-from gaphor.core.modeling import StyleSheet
+from gaphor.core.modeling import Presentation, StyleSheet
+from gaphor.diagram.group import self_and_owners
 from gaphor.core.modeling.diagram import Diagram, StyledDiagram
 from gaphor.core.modeling.event import AttributeUpdated, ElementDeleted
 from gaphor.diagram.diagramtoolbox import get_tool_def, tooliter
@@ -116,6 +118,7 @@ class DiagramPage:
         view = builder.get_object("view")
         view.selection = Selection()
         view.add_css_class(self._css_class())
+        view.connect("delete", delete_selected_items, self.event_manager)
 
         self.diagram_css = Gtk.CssProvider.new()
         Gtk.StyleContext.add_provider_for_display(
@@ -276,6 +279,17 @@ class DiagramPage:
         )
 
         view.request_update(self.diagram.get_all_items())
+
+
+def delete_selected_items(view: GtkView, event_manager):
+    with Transaction(event_manager):
+        items = view.selection.selected_items
+        for i in list(items):
+            assert isinstance(i, Presentation)
+            assert isinstance(i.diagram, UML.Diagram)
+            if i.subject and i.subject in self_and_owners(i.diagram):
+                del i.diagram.element
+            i.unlink()
 
 
 def context_menu_controller(context_menu, diagram):
