+++ gaphor/ui/diagramview.py
@@ -26,23 +26,9 @@ class DiagramView(GtkView):
     def _paste_full_clipboard(self):
         log.info("paste-full-clipboard")
 
-
-def _new_signal_shortcut(shortcut, signal):
-    return Gtk.Shortcut.new(
-        trigger=Gtk.ShortcutTrigger.parse_string(shortcut),
-        action=Gtk.SignalAction.new(signal),
-    )
-
-
-_mod = "<Meta>" if sys.platform == "darwin" else "<Control>"
-
-DiagramView.add_shortcut(_new_signal_shortcut(f"{_mod}x", "cut-clipboard"))
-DiagramView.add_shortcut(_new_signal_shortcut(f"{_mod}c", "copy-clipboard"))
-DiagramView.add_shortcut(_new_signal_shortcut(f"{_mod}v", "paste-clipboard"))
-DiagramView.add_shortcut(
-    _new_signal_shortcut(f"{_mod}<Shift>v", "paste-full-clipboard")
-)
-del _mod
+    @GObject.Signal(name="delete", flags=GObject.SignalFlags.RUN_LAST)
+    def _delete(self):
+        pass
 
 
 def _trigger_signal(signal_name):
@@ -58,3 +44,23 @@ DiagramView.install_action("clipboard.paste", None, _trigger_signal("paste-clipb
 DiagramView.install_action(
     "clipboard.paste-full", None, _trigger_signal("paste-full-clipboard")
 )
+DiagramView.install_action("diagram.delete", None, _trigger_signal("delete"))
+
+
+def _new_named_shortcut(shortcut, action_name):
+    return Gtk.Shortcut.new(
+        trigger=Gtk.ShortcutTrigger.parse_string(shortcut),
+        action=Gtk.NamedAction.new(action_name),
+    )
+
+
+_mod = "<Meta>" if sys.platform == "darwin" else "<Control>"
+
+DiagramView.add_shortcut(_new_named_shortcut(f"{_mod}x", "clipboard.cut"))
+DiagramView.add_shortcut(_new_named_shortcut(f"{_mod}c", "clipboard.copy"))
+DiagramView.add_shortcut(_new_named_shortcut(f"{_mod}v", "clipboard.paste"))
+DiagramView.add_shortcut(_new_named_shortcut(f"{_mod}<Shift>v", "clipboard.paste-full"))
+DiagramView.add_shortcut(
+    _new_named_shortcut("Delete|BackSpace|<Meta>BackSpace", "diagram.delete")
+)
+del _mod
