+++ pyproject.toml
@@ -122,7 +122,6 @@ gaphor = "gaphor.main:main"
 "main_window" = "gaphor.ui.mainwindow:MainWindow"
 "export_menu" = "gaphor.ui.menufragment:MenuFragment"
 "tools_menu" = "gaphor.ui.menufragment:MenuFragment"
-"copy" = "gaphor.ui.copyservice:CopyService"
 "sanitizer" = "gaphor.UML.sanitizerservice:SanitizerService"
 "recovery" = "gaphor.storage.recovery:Recovery"
 "model_browser" = "gaphor.ui.modelbrowser:ModelBrowser"
